Corregido algo de la documentación. Añadidas mas clases y daos y corregidas las anteriores.

+++ src/main/java/es/classone/restaurant/model/currency/CurrencyDao.java
@@ -4,6 +4,6 @@ import es.classone.restaurant.modelutil.dao.GenericDao;
 
 
 
-public interface CurrencyDao extends GenericDao<Currency, Long>{
+public interface CurrencyDao extends GenericDao<Currency, Integer>{
 
 }
