+++ src/main/java/es/classone/restaurant/model/currency/CurrencyDaoHibernate.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import es.classone.restaurant.modelutil.dao.GenericDaoHibernate;
 
 @Repository("currencyDao")
-public class CurrencyDaoHibernate extends GenericDaoHibernate<Currency, Long>
+public class CurrencyDaoHibernate extends GenericDaoHibernate<Currency, Integer>
 		implements CurrencyDao {
 
 }
