+++ src/main/java/es/classone/restaurant/model/dishGroup/DishGroupDao.java
@@ -2,6 +2,6 @@ package es.classone.restaurant.model.dishGroup;
 
 import es.classone.restaurant.modelutil.dao.GenericDao;
 
-public interface DishGroupDao extends GenericDao<DishGroup,Long>{
+public interface DishGroupDao extends GenericDao<DishGroup, Integer> {
 
 }
