+++ src/main/java/es/classone/restaurant/model/dishGroup/DishGroupDaoHibernate.java
@@ -4,9 +4,8 @@ import org.springframework.stereotype.Repository;
 
 import es.classone.restaurant.modelutil.dao.GenericDaoHibernate;
 
-
 @Repository("dishGroupDao")
 public class DishGroupDaoHibernate extends
-		GenericDaoHibernate<DishGroup, Long> implements DishGroupDao {
+		GenericDaoHibernate<DishGroup, Integer> implements DishGroupDao {
 
 }
