+++ src/main/java/es/classone/restaurant/model/methodOfPayment/MethodOfPaymentDao.java
@@ -4,6 +4,6 @@ import es.classone.restaurant.modelutil.dao.GenericDao;
 
 
 
-public interface MethodOfPaymentDao extends GenericDao<MethodOfPayment, Long> {
+public interface MethodOfPaymentDao extends GenericDao<MethodOfPayment, Integer> {
 
 }
