+++ src/main/java/es/classone/restaurant/model/qualifier/QualifierDao.java
@@ -4,6 +4,6 @@ import es.classone.restaurant.modelutil.dao.GenericDao;
 
 
 
-public interface QualifierDao extends GenericDao<Qualifier, Long> {
+public interface QualifierDao extends GenericDao<Qualifier, Integer> {
 
 }
