+++ src/main/java/es/classone/restaurant/model/waiter/WaiterDao.java
@@ -2,8 +2,6 @@ package es.classone.restaurant.model.waiter;
 
 import es.classone.restaurant.modelutil.dao.GenericDao;
 
-
-
-public interface WaiterDao extends GenericDao<Waiter, Long>{
+public interface WaiterDao extends GenericDao<Waiter, Integer> {
 
 }
