+++ src/main/java/es/classone/restaurant/model/waiter/WaiterDaoHibernate.java
@@ -2,6 +2,7 @@ package es.classone.restaurant.model.waiter;
 
 import es.classone.restaurant.modelutil.dao.GenericDaoHibernate;
 
-public class WaiterDaoHibernate extends GenericDaoHibernate<Waiter, Long> implements WaiterDao{
+public class WaiterDaoHibernate extends GenericDaoHibernate<Waiter, Integer>
+		implements WaiterDao {
 
 }
